Move plotting part of phantom.py to demo folder and update phantom.rst
