+++ demo/demo_utils.py
@@ -1,0 +1,25 @@
+import matplotlib.pyplot as plt
+
+def plot_image(img, title=None, filename=None, vmin=None, vmax=None):
+    """
+    Function to display and save a 2D array as an image.
+
+    Args:
+        img: 2D numpy array to display
+        title: Title of plot image
+        filename: A path to save plot image
+        vmin: Value mapped to black
+        vmax: Value mapped to white
+    """
+
+    plt.ion()
+    fig = plt.figure()
+    imgplot = plt.imshow(img, vmin=vmin, vmax=vmax)
+    plt.title(label=title)
+    imgplot.set_cmap('gray')
+    plt.colorbar()
+    if filename != None:
+        try:
+            plt.savefig(filename)
+        except:
+            print("plot_image() Warning: Can't write to file {}".format(filename))
